Moving custom field references to a general place
